requirements.txt
# MILESTONE 1 - CORE ML DEPENDENCIES
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# UI DEPENDENCIES 
streamlit>=1.25.0
plotly>=5.15.0

# HOSTING DEPENDENCIES
python-dotenv>=1.0.0

# MILESTONE 2 DEPENDENCIES (for later)
# langgraph>=0.2.0
# langchain>=0.3.0
# chromadb>=1.4.0
# openai>=1.0.0
# faiss-cpu>=1.7.0
